Enhancement/193/workstation scripts update (#98)
* get user by upi

* don't lock certain user types

* ignore user types when unlocking account

* ignore user types when deleteing an account

* dont create user if user exists

* refactoring

* lock workstation user after account creation

* removed other from ignore

* rename to shouldUserAccountLockOrExpire

* renamed to shouldPerformUserOpertaions

+++ backend/src/ssh/deleteWorkstationUser.js
@@ -1,11 +1,17 @@
 import { SSHError } from 'node-ssh';
 import { execCommand } from '.';
+import { shouldPerformUserOperation } from './util/shouldPerformUserOperation';
 
 /**
  * @param {string} host ip address of machine to ssh into
  * @param {string} upi the account to be deleted
  */
 export async function deleteWorkstationUser(host, upi) {
+    const shouldDelete = await shouldPerformUserOperation(upi);
+    if (!shouldDelete) {
+        return;
+    }
+
     // Deletes the user, their home directory and mail spool.
     const result = await execCommand(host, `sudo userdel -r ${upi}`);
